refactor!: move `LuaVersion` out of `config` module
