+++ lux-cli/src/install.rs
@@ -1,10 +1,7 @@
 use eyre::Result;
 use lux_lib::{
-    config::{Config, LuaVersion},
-    lockfile::PinnedState,
-    operations,
-    package::PackageReq,
-    progress::MultiProgress,
+    config::Config, lockfile::PinnedState, lua_version::LuaVersion, operations,
+    package::PackageReq, progress::MultiProgress,
 };
 
 use crate::utils::install::apply_build_behaviour;
