+++ lux-cli/src/install_lua.rs
@@ -1,7 +1,8 @@
 use eyre::{OptionExt, Result};
 use lux_lib::{
-    config::{Config, LuaVersion},
+    config::Config,
     lua_installation::LuaInstallation,
+    lua_version::LuaVersion,
     progress::{MultiProgress, ProgressBar},
 };
