+++ lux-cli/src/pin.rs
@@ -3,8 +3,9 @@ use eyre::eyre;
 use eyre::Context;
 use eyre::Result;
 use itertools::Itertools;
-use lux_lib::config::{Config, LuaVersion};
+use lux_lib::config::Config;
 use lux_lib::lockfile::PinnedState;
+use lux_lib::lua_version::LuaVersion;
 use lux_lib::operations;
 use lux_lib::package::PackageName;
 use lux_lib::package::PackageReq;
