+++ lux-lib/src/lockfile/mod.rs
@@ -1363,10 +1363,7 @@ mod tests {
     use assert_fs::fixture::PathCopy;
     use insta::{assert_json_snapshot, sorted_redaction};
 
-    use crate::{
-        config::{ConfigBuilder, LuaVersion::Lua51},
-        package::PackageSpec,
-    };
+    use crate::{config::ConfigBuilder, lua_version::LuaVersion, package::PackageSpec};
 
     #[test]
     fn parse_lockfile() {
@@ -1382,7 +1379,7 @@ mod tests {
             .user_tree(Some(temp.to_path_buf()))
             .build()
             .unwrap();
-        let tree = config.user_tree(Lua51).unwrap();
+        let tree = config.user_tree(LuaVersion::Lua51).unwrap();
         let lockfile = tree.lockfile().unwrap();
 
         assert_json_snapshot!(lockfile, { ".**" => sorted_redaction() });
@@ -1411,7 +1408,7 @@ mod tests {
             .user_tree(Some(temp.to_path_buf()))
             .build()
             .unwrap();
-        let tree = config.user_tree(Lua51).unwrap();
+        let tree = config.user_tree(LuaVersion::Lua51).unwrap();
         let mut lockfile = tree.lockfile().unwrap().write_guard();
 
         let test_package = PackageSpec::parse("test1".to_string(), "0.1.0".to_string()).unwrap();
@@ -1456,7 +1453,7 @@ mod tests {
             .user_tree(Some(temp.to_path_buf()))
             .build()
             .unwrap();
-        let tree = config.user_tree(Lua51).unwrap();
+        let tree = config.user_tree(LuaVersion::Lua51).unwrap();
 
         let _ = tree.lockfile().unwrap().write_guard(); // Try to create the lockfile but don't actually do anything with it.
     }
@@ -1610,7 +1607,7 @@ mod tests {
             .user_tree(Some(temp.to_path_buf()))
             .build()
             .unwrap();
-        let tree = config.user_tree(Lua51).unwrap();
+        let tree = config.user_tree(LuaVersion::Lua51).unwrap();
         let lockfile = tree.lockfile().unwrap();
 
         let packages: Vec<LuaDependencySpec> = vec![
